Modularize geometry filling.
