+++ RecHitAnalyzer/plugins/BuildFile.xml
@@ -3,6 +3,7 @@
 <use name="FWCore/ParameterSet"/>
 <flags EDM_PLUGIN="1"/>
  <use name="DataFormats/TrackReco"/>
+ <use name="DataFormats/TrackingRecHit"/>
  <use name="DataFormats/EcalRecHit"/>
  <use name="DataFormats/HcalDetId"/>
  <use name="DataFormats/HcalRecHit"/>
